Rename package to redactwriter
